Moves contract-client-gen logic to PublicImpl
