+++ stylus-proc/src/macros/public/types.rs
@@ -9,6 +9,7 @@ use syn::{
 };
 
 use crate::{
+    consts::STYLUS_CONTRACT_ADDRESS_FIELD,
     imports::{
         alloy_sol_types::SolType,
         stylus_sdk::abi::{AbiType, Router},
@@ -185,6 +186,81 @@ impl PublicImpl {
                 }
             }
         })
+    }
+
+    pub fn contract_client_gen(&self) -> proc_macro2::TokenStream {
+        let client_funcs = self
+        .funcs
+        .iter()
+        .map(|func| {
+            let func_name = func.name.clone();
+
+            let (context, call) = func.purity.get_context_and_call();
+
+            let inputs = func.inputs.iter().map(|input| {
+                let name = input.name.clone();
+                let ty = input.ty.clone();
+                quote! { #name: #ty }
+            });
+            let inputs_names = func.inputs.iter().map(|input| {
+                input.name.clone()
+            });
+            let inputs_types = func.inputs.iter().map(|input| {
+                let ty = input.ty.clone();
+                quote! { #ty }
+            });
+
+            let output_type = match &func.output {
+                syn::ReturnType::Type(_, ty) => {
+                    let ty = ty.clone();
+                    quote! { #ty }
+                }
+                syn::ReturnType::Default => {
+                    quote! { () }
+                }
+            };
+
+            let function_selector = func.function_selector();
+
+            quote! {
+                pub fn #func_name(
+                    &self,
+                    host: &dyn stylus_sdk::stylus_core::host::Host,
+                    context: impl #context,
+                    #(#inputs,)*
+                ) -> Result<<<#output_type as #AbiType>::SolType as #SolType>::RustType, stylus_sdk::stylus_core::calls::errors::Error> {
+                    let inputs = <<(#(#inputs_types,)*) as #AbiType>::SolType as #SolType>::abi_encode_params(&(#(#inputs_names,)*));
+                    use stylus_sdk::function_selector;
+                    let mut calldata = #function_selector;
+                    let call_result = #call(host, context, self.#STYLUS_CONTRACT_ADDRESS_FIELD, &calldata)?;
+                    Ok(<<#output_type as #AbiType>::SolType as #SolType>::abi_decode(&call_result)?)
+                }
+            }
+        })
+        .collect::<proc_macro2::TokenStream>();
+
+        let struct_path = self.self_ty.clone();
+
+        let mut output = quote! {
+            impl #struct_path {
+                #client_funcs
+            }
+        };
+
+        // If the impl does not implement a trait, we add a constructor for the contract client
+        if self.trait_.is_none() {
+            output.extend(quote! {
+                impl #struct_path {
+                    pub fn new(address: stylus_sdk::alloy_primitives::Address) -> Self {
+                        Self {
+                            #STYLUS_CONTRACT_ADDRESS_FIELD: address,
+                        }
+                    }
+                }
+            });
+        }
+
+        output.into()
     }
 }
